todos save in sqlite db and have sql injection protection also some validation added to todos apis

+++ src/controllers/TodoController.ts
@@ -9,9 +9,14 @@ export default class TodoController{
         
         app.post(this.prePath + "/add", (req, res) => {
             
+            if (req.body.name == undefined || req.body.description == undefined) {
+                res.status(400).send("send name and description in body")
+                return
+            }
+
             let todo = Todo.getInstance(
-                req.header('name'),
-                req.header('description')
+                req.body.name,
+                req.body.description
             )
             TodoRepo.add(todo)
             res.status(200).send("done")
@@ -19,17 +24,27 @@ export default class TodoController{
 
         app.post(this.prePath + "/remove", (req, res) => {
 
+            if (req.header('todoId') == undefined) {
+                res.status(400).send("set todoid in header")
+                return
+            }
+
             TodoRepo.delete(req.header('todoId'))
             res.status(200).send("done")
         })
 
         app.post(this.prePath + "/mark", (req, res) => {
 
+            if (req.header('todoId') == undefined) {
+                res.status(400).send("set todoid in header")
+                return
+            }
+
             TodoRepo.mark(req.header('todoId'))
             res.status(200).send("done")
         })
 
-        app.post(this.prePath + "/getAll", (req, res) => {
+        app.get(this.prePath + "/getAll", (req, res) => {
 
             TodoRepo.getAll((todos: Todo[]) => {
